Release 2.7 version

+++ threads/FFmpegInstallThread.py
@@ -1,4 +1,5 @@
 import os
+import re
 import shutil
 import subprocess
 import traceback
@@ -32,17 +33,32 @@ class FFmpegInstallThread(QThread):
                 ["ffmpeginstall.bat"],
                 cwd=os.getcwd(),
                 stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE
+                stderr=subprocess.STDOUT,
+                text=True,  # Декодируем вывод в текст (Python 3.7+)
+                bufsize=1,  # Построчная буферизация
+                universal_newlines=True,  # Аналогично text=True, для совместимости
+                encoding="utf-8"
             )
 
-            # Чтение вывода процесса
-            while self.process.poll() is None:
-                if self.cancel_install:
-                    self.process.terminate()
-                    self.progress_signal.emit("Установка отменена.")
-                    self.finished_signal.emit(False)
-                    return
+            if self.process.stdout != None:
+                for line in self.process.stdout:
+                    line = line.strip()
+                    if line:
+                        match = re.match(r"(\d+)", line)  # Ищем число в начале строки
+                        if match:
+                            percent = int(match.group(1))  # Преобразуем в int
+                            self.progress_signal.emit(f"Скачивание FFmpeg... {percent}%")
+                        self.config.log('FFmpegInstallThread', 'run', line)
+                    
+                    if self.cancel_install:
+                        if hasattr(self, "process") and self.process:
+                            self.process.terminate()
+                        self.progress_signal.emit("Установка отменена.")
+                        self.finished_signal.emit(False)
+                        return
 
+            # Дождаться завершения процесса
+            self.process.wait()
             self.progress_signal.emit("Установка завершена!")
             self.finished_signal.emit(True)
 
@@ -51,10 +67,8 @@ class FFmpegInstallThread(QThread):
 
     def handle_exception(self, e):
         error_message = ''.join(traceback.format_exception(type(e), e, e.__traceback__))
-        self.config.log('FFmpegInstallThread', f"Error: {error_message}")
+        self.config.log('FFmpegInstallThread', 'handle_exception', f"Error: {error_message}")
         self.error_signal.emit(e)
 
     def cancel(self):
         self.cancel_install = True
-        if hasattr(self, "process") and self.process:
-            self.process.terminate()
